MachineLearning.Helpers

Documentation

unlines' :: [String] -> String

Unlines variant which does not add a newline after the last element.

chunks :: [a] -> [Int] -> [[a]]

Chops a list into chunks of predefined lengths.

>>> chunks (repeat 42) [1, 2, 1, 3]
[[42], [42, 42], [42], [42, 42, 42]]

chunks' :: [a] -> Int -> [[a]]

Chops a list into equally sized chunks.

>>> chunks' [42, 42, 42, 42, 42] 2
[[42, 42], [42, 42], [42]]

showFFloat' :: Int -> Float -> String

Shows a Float with the least number of decimals possible, with a maximum of d decimals.

>>> showFFloat' 3 2.0000001
"2"
>>> showFFloat' 3 2.1234567
"2.123"
>>> showFFloat' 3 10.000001
"10"

showFDouble' :: Int -> Double -> String