MachineLearning.LinearRegression.LinearRegression

Minimization

addX0 :: Matrix Double -> Matrix Double

Adds a column with 1 values at the front of a matrix.

hypothesis :: Matrix Double -> Matrix Double -> Matrix Double

Hypothesis function.

 h(theta) = X * theta T

cost :: Matrix Double -> Matrix Double -> Matrix Double -> Double

Cost function

 J(theta) = 1/2m * sum( (h(theta) - y) ^ 2)

cost' :: Matrix Double -> Matrix Double -> Matrix Double -> Matrix Double

Partial derivative of the cost function.

 J(theta)' = 1/m * sum ( h(theta) - y ) 

minimize

Arguments

:: Matrix Double
Feature matrix.

-> Matrix Double
Expected outcomes (column-vector).

-> Matrix Double
Theta values (row-vector)

-> Double
Learning rate.

-> Int
Maximum number of iterations.

-> Matrix Double

Minimizes J(theta) over a training set.

Helpers

dataFromFile :: FilePath -> IO (Matrix Double, Matrix Double)

Reads a set of feature values from a csv file.

runCost :: FilePath -> Matrix Double -> IO Double

run

Arguments

:: FilePath
Features csv file.

-> Double
Learning rate.

-> Int
Maximum number of iterations.

-> IO (Matrix Double)

Runs the minimisation algorithm on data read from a file.