MachineLearning.Maths.Matrix

Description

A simple matrix module which supports basic matrix transformations such as by-element and scalar arithmetic operations and matrix multiplication.

Constructors

data Matrix a

Constructors

M

Fields

rows :: [[a]]

(<+>) :: (Int, Int) -> [a] -> Matrix a

Block-matrix constructor. The first value in the tuple is the number of rows, the second the number of columns.

>>> (2, 3) <+> [1..]
[1, 2, 3]
[4, 5, 6]

(<|>) :: Int -> [a] -> Matrix a

Column vector constructor.

>>> 2 <|> [1..]
[1]
[2]

(<->) :: Int -> [a] -> Matrix a

Row vector constructor.

>>> 3 <-> [1..]
[1, 2, 3]

(<\>) :: (Int, Int) -> (a, a) -> Matrix a

Identity matrix constructor.

>>> (2, 3) <\> ('x', 'y')
['y', 'x', 'x']
['x', 'y', 'x']

mIdentity :: Num a => Int -> Matrix a

Creates a square numeric identity matrix

>>> mIdentity 2 :: Matrix Float
[1.0, 0.0]
[0.0, 1.0]

mEmpty :: Matrix a

Creates a matrix with 0 elements.

mInfinite :: a -> Matrix a

Creates a matrix with an infinite size.

Matrix Properties

numRows :: Matrix a -> Int

Gets the number of rows in a matrix.

numCols :: Matrix a -> Int

Gets the number of columns in a matrix.

numElements :: Matrix a -> Int

Gets the number of cells in a matrix.

size :: Matrix a -> (Int, Int)

Gets the size of a matrix.

isEmpty :: Eq a => Matrix a -> Bool

Tests if the matrix is empty.

Sub-matrices

cols :: Matrix a -> [[a]]

Gets a list of columns of a matrix.

elements :: Matrix a -> [a]

Gets a single list with all elements of a matrix.

element :: Matrix a -> (Int, Int) -> a

Gets the element in row r, column c.

row :: Matrix a -> Int -> Matrix a

Gets the i-th row of a matrix.

col :: Matrix a -> Int -> Matrix a

Gets the i-th column of a matrix.

(!!!) :: Matrix a -> (Int, Int) -> a

Alternative to the element function.

(!-!) :: Matrix a -> Int -> Matrix a

Alternative to the row function.

(!|!) :: Matrix a -> Int -> Matrix a

Alternative to the col function.

mFirst :: Matrix a -> a

Returns the upper-left value of the matrix.

(+-+) :: Matrix a -> Matrix a -> Matrix a

Appends matrix b to matrix a as new rows.

(+|+) :: Matrix a -> Matrix a -> Matrix a

Appends matrix b to matrix a as new columns.

takeR :: Int -> Matrix a -> Matrix a

Takes n rows from a matrix.

dropR :: Int -> Matrix a -> Matrix a

Drops n rows from a matrix.

takeC :: Int -> Matrix a -> Matrix a

Takes n columns from a matrix.

dropC :: Int -> Matrix a -> Matrix a

Drops n rows from a matrix.

repeatR :: Matrix a -> Matrix a

Repeats the rows of a matrix

repeatC :: Matrix a -> Matrix a

Repeats the columns of a matrix

Matrix Transformations

mTranspose :: Matrix a -> Matrix a

Transpose the matrix.

mMap :: (a -> b) -> Matrix a -> Matrix b

Maps a function over each element in the matrix.

rMap :: ([a] -> [b]) -> Matrix a -> Matrix b

Maps a function over each row in the matrix.

cMap :: ([a] -> [b]) -> Matrix a -> Matrix b

Maps a function over each column in the matrix.

mZip :: (a -> b -> c) -> Matrix a -> Matrix b -> Matrix c

Zips two matrices with a given function f.

Arithmetic operations

(!+) :: Num a => Matrix a -> a -> Matrix a

Scalar addition.

(!-) :: Num a => Matrix a -> a -> Matrix a

Scalar subtraction.

(!*) :: Num a => Matrix a -> a -> Matrix a

Scalar multiplication.

(!/) :: Fractional a => Matrix a -> a -> Matrix a

Scalar division.

mMul :: Num a => Matrix a -> Matrix a -> Matrix a

Matrix multiplication.

Miscellaneous functions

chunksC' :: Matrix a -> Int -> [Matrix a]

Chops a matrix into column chunks of the given length.

>>> chunksC' ((2, 6) <+> [1..]) 3
[[1,2,3]
 [7,8,9],
 [4,5,6]
 [10,11,12]]

m1 :: Matrix Integer

v1 :: Matrix Integer

m2 :: Matrix Integer

v2 :: Matrix Integer

m3 :: Matrix Integer

i1 :: Matrix Integer