MachineLearning.Maths.Statistics

Description

A collection of statistical operations on matrices.

Sum and mean values

rSum :: Num a => Matrix a -> Matrix a

Sums up all rows and returns the sums as a column vector

cSum :: Num a => Matrix a -> Matrix a

Sums up all columns and returns the sums as a row vector

mSum :: Num a => Matrix a -> a

Sums up all values in a matrix.

rMean :: Floating a => Matrix a -> Matrix a

Calculates the mean value of each row in a matrix.

cMean :: Floating a => Matrix a -> Matrix a

Calculates the mean value of each column in a matrix.

mMean :: Floating a => Matrix a -> a

Calculates the mean over of values in a matrix.

Quantile

quantile :: RealFrac a => a -> [a] -> a

Gets a given quantile from a list of values.

>>> quantile 0.25 [1,2,3,4,5]
2.0
>>> quantile 0.75 [1,2,3,4,5]
4.0

median :: RealFrac a => [a] -> a

Gets the median from a list of values.

quartiles :: RealFrac a => [a] -> (a, a, a)

Gets three values dividing a list into quartiles. (Q1, median, Q3).

>>> quartiles [1,2,3,4]
(1.5,2.5,3.5)

fiveNum :: RealFrac a => [a] -> (a, a, a, a, a)

Gets the five-number summary from a list. (min, Q1, median, Q3, max).

>>> fiveNum [1,2,3,4]
(1.0,1.5,2.5,3.5,4.0)

Standard Deviation and Normalization

rStdDev :: Floating a => Matrix a -> Matrix a

Returns a column vector with the standard deviation of each row in a matrix.

cStdDev :: Floating a => Matrix a -> Matrix a

Returns a row vector with the standard deviation over all rows.

rNorm :: Floating a => Matrix a -> (Matrix a, Matrix a, Matrix a)

Normalizes each row in a matrix.

Returns a triple: (Matrix normalized over columns, means column vector, std. dev. column vector).

cNorm :: Floating a => Matrix a -> (Matrix a, Matrix a, Matrix a)

Normalizes each column in a matrix.

Mean Squared Error

mse :: Floating a => Matrix a -> Matrix a -> a

Mean squared error.

 mse(x,y) = 1/2n * sum( (x(i) - y(i)) ^ 2)

rmse :: Floating a => Matrix a -> Matrix a -> a

Root mean squared error.

Correlations

corr :: Floating a => Matrix a -> Int -> Int -> a

Pearson's correlation between columns x and y in matrix m.

 corr(X,Y) = cov(X,Y) / (stdDev(X) stdDev(Y))

covariance :: Floating a => Matrix a -> Matrix a -> a

Covariance between two column-vectors.

 q(jk) = 1 / (n - 1) * sum (x(ij) - mean( x(j) )) * ( x(ik) - mean( x(j) ) )