MachineLearning.Requisitions.IO

Read

trainingSetFromCSV

Arguments

:: String
CSV input string.

-> Bool
Flip polarities

-> (FeatureSet -> FeatureSet)
Transformation function.

-> IO (TrainingSet, Matrix String)

Reads a set of training examples from a CSV string.

All lines will be grouped into matrices by their first value, REQ_ID. At the moment the first occurrence of a new REQ_ID is assumed to be the picked supplier.

Expected input format:

 REQ_ID, SUPPLIER_ID, x1, x2, x3
 REQ_ID, SUPPLIER_ID, x1, x2, x3

trainingSetFromCSV'

Arguments

:: FilePath
The csv file to read.

-> Bool
Flip polarities

-> (FeatureSet -> FeatureSet)
Transformation function.

-> IO (TrainingSet, Matrix String)

Reads a set of training examples from a CSV file.

Write

trainingSetToCSV :: TrainingSet -> String

Writes a training set to a comma separated values string. Well..semicolon separated really...

trainingSetToCSV' :: FilePath -> TrainingSet -> IO ()

Writes a training set to a csv file.