MachineLearning.Requisitions.Random

Default values

defaultFMin :: Int

The default minimum feature value.

defaultFMax :: Int

The default maximum feature value.

Generating new sets

generateSet

Arguments

:: Int
The number of orders to generate.

-> Int
The number of suppliers per order.

-> Int
The number of features per supplier.

-> [Int]
The feature to base picking the best supplier on.

-> IO TrainingSet

Generates a random training set.

generateSet'

Arguments

:: Int
The number of orders to generate.

-> Int
The number of suppliers per order.

-> Int
The number of features per supplier.

-> Int
Minimal feature value.

-> Int
Maximum feature value.

-> [Int]
The feature to base picking the best supplier on.

-> IO TrainingSet

Generates a random training set.

generateFeatures

Arguments

:: Int
The number of suppliers to generate.

-> Int
The number of features per supplier.

-> IO FeatureSet

Generates a single random feature set.

generateFeatures'

Arguments

:: Int
The number of suppliers to generate.

-> Int
The number of features per supplier.

-> Int
Minimal feature value.

-> Int
Maximum feature value.

-> IO FeatureSet

Generates a single random feature set.

generatePicks

Arguments

:: [FeatureSet]
The features to compare.

-> [Int]
The feature column indices to base the picking on.

-> [Int]

Generates a list of picked supplier indices, based on the highest value in the given columns.

Inserting errors

insertErrors

Arguments

:: TrainingSet
The training set to modify.

-> Int
Number of errors to introduce.

-> IO TrainingSet

Changes the picked value in a given number of random training examples to a random value, where the new value /= original value.

Helpers

randRangeExcept :: Int -> Int -> [Int] -> IO Int

Generates a random integer in the supplied range. It ensures that the generated number is not an element in the supplied exclude list.

randOneOf :: [a] -> IO a

Picks a random element from the supplied list.

uniqueRandList

Arguments

:: Int
The lower range bound.

-> Int
The upper range bound.

-> Int
The number of numbers to generate.

-> IO [Int]

Generates a list of unique random numbers in the supplied range.