MachineLearning.Requisitions.Requisitions

Type Synonyms

type FeatureSet = Matrix Double

A block-matrix with supplier feature values (features across columns).

type FeatureWeights = Matrix Double

A row-vector of feature weights.

type TrainingExample = (FeatureSet, Int)

A combination of a feature matrix and picked supplier index.

type TrainingSet = [TrainingExample]

type RankFunction

Arguments

= FeatureSet
Set of supplier features (matrix).

-> FeatureWeights
Feature weights (row vector).

-> Matrix Double
Supplier scores (column vector).

Configurations

data AlgorithmConfig

Constructors

AlgorithmConfig

Fields

getTrainingFn :: TrainingSet -> FeatureWeights

getRankingFn :: RankFunction

getPreScaleFn :: FeatureSet -> FeatureSet

getPostScaleFn :: FeatureSet -> FeatureSet

getFlipPolarities :: Bool

getFeatureGroupSize :: Float

algorithmConfigs :: [(Char, AlgorithmConfig)]

getAlgorithmConfig :: Char -> AlgorithmConfig

featureLabels :: Float -> [String]

Functions

scaleFeatures :: FeatureSet -> FeatureSet

Scales the columns of the given matrix such that the values are all within the range <0, 1>.

featureWeightsDAvg :: TrainingSet -> FeatureWeights

Maps distanceToAvg function over all training examples and sums the results.

featureWeightsSum :: TrainingSet -> FeatureWeights

Alternative method of getting feature weights. Instead of distance to average feature value, it uses the absolute feature values as the weights.

rank :: RankFunction

Ranks the supplier's feature set, returning a column-vector of scores.

 r = features * (weights T)

pickBest :: RankFunction -> FeatureSet -> FeatureWeights -> Int

Returns the index of the highest-ranked supplier.

errorRate :: RankFunction -> TrainingSet -> FeatureWeights -> Double

Calculates the errorrate of the supplied weights when using them on a training set. Lower value is better.

flipPolarity :: FeatureSet -> FeatureSet

Flip the polarity of the entire feature set (negating all values)

flipPolarity' :: [Bool] -> FeatureSet -> FeatureSet

Flips the polarity of the columns in the supplied boolean list.

addPolarizedFeatures :: FeatureSet -> FeatureSet

Add polarized features to feature set. p -> p-, pAverage, p+

addPolarizedFeatures' :: FeatureSet -> FeatureSet

Add polarized features to feature set. p -> p-, p+

rank' :: RankFunction

Rank using only highest weight of the polarized features.

rank'' :: RankFunction

Rank using only highest weight of the + and - polarities. The resulting weight used is the (absolute) difference between + and -.